requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
Django==4.2.16
django-autoslug==1.9.9
django-ckeditor==6.7.3
django-ckeditor-5==0.2.19
django-js-asset==3.1.2
django-taggit==6.1.0
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.2
pillow==12.1.0
platformdirs==4.5.1
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.3.0
sqlparse==0.5.5
text-unidecode==1.3
dotenv
# ===== PROD =====
gunicorn==22.0.0
psycopg[binary]==3.2.3
